refactor(library): decompose LibraryService into fetcher/orchestrator/reporter
Apply the saves-vertical playbook to services/library.py (1123 LOC).
The single LibraryService class handled platform/collection fetching,
preview/apply lifecycle, full sync orchestration with safety
heartbeat, post-apply registry updates, and registry queries — five
distinct responsibilities in one place.

services/library/ now contains:
- _state.py: LibrarySyncStateBox — the mutable shared state
  (sync_state, current_sync_id, heartbeat, progress dict, pending_*
  dicts and the PreviewDelta).
- fetcher.py: LibraryFetcher — platform/collection metadata callables
  and the full ROM fetch pipeline (incremental skip, paginated full
  fetch, collection roundtrips, metadata-cache stamping). Owns
  _fetch_and_prepare.
- sync_orchestrator.py: SyncOrchestrator — preview/apply/full-sync
  lifecycle, _emit_progress, the safety-timeout watchdog, artwork
  download delegation.
- reporter.py: SyncReporter — report_sync_results, _build_collection_app_ids,
  _finalize_cover_path, registry queries (get_registry_platforms,
  get_sync_stats, get_rom_by_steam_app_id), clear_sync_cache.
- service.py: LibraryService façade — composes the three sub-services
  over the shared box and exposes the public callable surface
  consumed by main.py. Class body 344 LOC.

Each sub-service has its own *ServiceConfig (Always-Config rule);
sub-services receive Protocol-typed adapters plus a reference to the
shared LibrarySyncStateBox instead of a back-ref to the façade. The
orchestrator's _emit_progress is threaded into the fetcher and
reporter through a late-bound proxy on the façade, avoiding a
construction-order cycle.

Public API preserved: bootstrap.py still imports
LibraryService/LibraryServiceConfig from services.library (the
package's __init__.py re-exports them), and every callable on
main.py keeps the same signature.

tests/services/test_library.py moves to tests/services/library/test_service.py;
test bodies updated to reach sub-service methods directly
(svc._fetcher._fetch_and_prepare, svc._orchestrator._emit_progress,
svc._reporter._report_sync_results_io). State and dependency
rebinding (svc._sync_state = ..., svc._sleeper = ...) flows through
property shims on the façade so tests stay readable.

All 2237 tests pass; ruff, basedpyright, import-linter (7 contracts),
and the cosmic-call-bans script remain clean.
